refactor: change all elements used only in PunsForm to it
